Don't throw exception if trying to delete something that doesn't exist

+++ src/main/java/com/redhat/summit2019/springmusic/api/AlbumController.java
@@ -89,6 +89,8 @@ public class AlbumController {
 	@ResponseStatus(HttpStatus.NO_CONTENT)
 	public void deleteById(@Parameter(description = "The album id") @PathVariable String id) {
 		LOGGER.info("Deleting album {}", id);
-		this.repository.deleteById(id);
+		if (this.repository.existsById(id)) {
+			this.repository.deleteById(id);
+		}
 	}
 }
